+++ smoothing/c/utils/common.cpp
@@ -1,9 +1,9 @@
 #include "common.h"
 
 // mean = 0, var = 1
-double randn()
+float randn()
 {
-    static double x1, x2, w = 1.5;
+    static float x1, x2, w = 1.5;
     static bool which = 0;
 
     if(which == 1)
@@ -23,7 +23,7 @@ double randn()
     return x1 * w;
 };
 
-void multivar_normal(double *mean, double *var, double *ret, int dim)
+void multivar_normal(float *mean, float *var, float *ret, int dim)
 {
     for(int i=0; i < dim; i++)
         ret[i] = mean[i] + sqrt(var[i])*randn();
